requirements.txt
Django
Pillow
pytz==2018.6
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
category==0.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
frozenlist==1.3.1
geoip2==4.6.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
maxminddb==2.2.0
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
phonenumbers==8.12.54
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2018.6
requests==2.28.1
requests-oauthlib==1.3.1
rust_category==0.2.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
yarl==1.8.1
